requirements.txt
asgiref==3.6.0
boto3==1.26.122
botocore==1.29.122
coverage==7.2.4
dj-database-url==0.5.0
Django==3.2.18
django-storages==1.13.2
djangorestframework==3.14.0
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
s3transfer==0.6.0
sqlparse==0.4.4
